+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Adding functionality to Run/Reset an indexer
 - Renaming the search feature in the dropdown from "Search Index" to "Query Index" for clarity
+- Adding support for queries in POST format
 
 ## 0.1.0 - 2020-09-16